refactor: 도서관 step 이동
