+++ src/main/java/com/bookspot/batch/library/reader/api/LibraryItemReader.java
@@ -1,7 +1,6 @@
 package com.bookspot.batch.library.reader.api;
 
-import com.bookspot.batch.library.LibraryStepConst;
-import com.bookspot.batch.library.reader.api.LibraryApiRequester;
+import com.bookspot.batch.step.LibraryStepConst;
 import com.bookspot.batch.library.data.Library;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
