+++ src/main/java/com/bookspot/batch/library/reader/crawler/NaruDetailParser.java
@@ -4,7 +4,7 @@ import com.bookspot.batch.global.crawler.common.CrawlingResult;
 import com.bookspot.batch.global.crawler.common.JsoupCrawler;
 import com.bookspot.batch.global.crawler.common.exception.ElementNotFoundException;
 import com.bookspot.batch.global.crawler.naru.NaruPagingUrlBuilder;
-import com.bookspot.batch.library.LibraryStepConst;
+import com.bookspot.batch.step.LibraryStepConst;
 import com.bookspot.batch.library.data.LibraryNaruDetail;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
